Move datafile package outside hive
And place it at the same level
